wip break apart into multi-stage apis for decoding
